refactor: extract project sidebar lists
