+++ frontend/src/components/project/ReferenceList.vue
@@ -1,0 +1,70 @@
+<template>
+  <q-expansion-item
+    v-model="expandedModel"
+    dense
+    expand-separator
+    class="text-grey-7"
+    style="min-height: 50px;"
+  >
+    <template #header>
+      <q-item-section class="text-caption text-weight-medium text-uppercase" style="letter-spacing: 0.05em;">
+        References ({{ references.length }})
+      </q-item-section>
+      <q-item-section side>
+        <BaseButton size="sm" variant="outline" @click.stop="emit('add')">
+          Add Reference
+        </BaseButton>
+      </q-item-section>
+    </template>
+
+    <div class="q-pa-none" style="max-height: 128px; overflow-y: auto;">
+      <q-list dense class="q-pa-none">
+        <q-item
+          v-for="reference in references"
+          :key="reference.id"
+          dense
+          clickable
+          class="q-pa-xs rounded hover:bg-grey-2 cursor-pointer"
+          @click="emit('select', reference)"
+        >
+          <q-item-section avatar class="min-width-auto q-pr-xs">
+            <q-icon name="description" size="14px" color="red-5" />
+          </q-item-section>
+          <q-item-section>
+            <q-item-label class="text-caption text-weight-medium text-grey-9" lines="1">
+              {{ reference.filename || reference.title }}
+            </q-item-label>
+            <q-item-label caption class="text-grey-6">
+              {{ reference.filetype || 'Unknown' }} • {{ formatFileSize(reference.filesize) }}
+            </q-item-label>
+          </q-item-section>
+        </q-item>
+      </q-list>
+    </div>
+  </q-expansion-item>
+</template>
+
+<script setup>
+import { computed } from 'vue'
+import BaseButton from '../ui/BaseButton.vue'
+
+const props = defineProps({
+  references: { type: Array, default: () => [] },
+  expanded: { type: Boolean, default: true }
+})
+
+const emit = defineEmits(['update:expanded', 'select', 'add'])
+
+const expandedModel = computed({
+  get: () => props.expanded,
+  set: val => emit('update:expanded', val)
+})
+
+function formatFileSize(bytes) {
+  if (bytes === 0) return '0 B'
+  if (!bytes) return ''
+  const sizes = ['B', 'KB', 'MB', 'GB', 'TB']
+  const i = Math.floor(Math.log(bytes) / Math.log(1024))
+  return `${(bytes / Math.pow(1024, i)).toFixed(1)} ${sizes[i]}`
+}
+</script>
