+++ frontend/src/pages/ProjectPage.vue
@@ -77,94 +77,17 @@
           <!-- References and Media Files List -->
           <div class="col">
             <q-list dense class="q-pa-none">
-              <!-- References section -->
-              <q-expansion-item
-                v-model="referencesExpanded"
-                dense
-                expand-separator
-                class="text-grey-7"
-                style="min-height: 50px;"
-              >
-                <template v-slot:header>
-                  <q-item-section class="text-caption text-weight-medium text-uppercase" style="letter-spacing: 0.05em;">
-                    References ({{ references.length }})
-                  </q-item-section>
-                  <q-item-section side>
-                    <BaseButton
-                      size="sm"
-                      variant="outline"
-                      @click.stop="showReferenceSearch = true"
-                    >
-                      Add Reference
-                    </BaseButton>
-                  </q-item-section>
-                </template>
-                
-                <div class="q-pa-none" style="max-height: 128px; overflow-y: auto;">
-                  <q-list dense class="q-pa-none">
-                    <q-item
-                      v-for="reference in references"
-                      :key="reference.id"
-                      dense
-                      clickable
-                      class="q-pa-xs rounded hover:bg-grey-2 cursor-pointer"
-                      @click="openReference(reference)"
-                    >
-                      <q-item-section avatar class="min-width-auto q-pr-xs">
-                        <q-icon name="description" size="14px" color="red-5" />
-                      </q-item-section>
-                      <q-item-section>
-                        <q-item-label class="text-caption text-weight-medium text-grey-9" lines="1">
-                          {{ reference.filename || reference.title }}
-                        </q-item-label>
-                        <q-item-label caption class="text-grey-6">
-                          {{ reference.filetype || 'Unknown' }} • {{ formatFileSize(reference.filesize) }}
-                        </q-item-label>
-                      </q-item-section>
-                    </q-item>
-                  </q-list>
-                </div>
-              </q-expansion-item>
-
-              <!-- Media Files section -->
-              <q-expansion-item
-                v-model="mediaFilesExpanded"
-                dense
-                expand-separator
-                class="text-grey-7"
-                style="min-height: 50px;"
-              >
-                <template v-slot:header>
-                  <q-item-section class="text-caption text-weight-medium text-uppercase" style="letter-spacing: 0.05em;">
-                    Media Files ({{ mediaFiles.length }})
-                  </q-item-section>
-                </template>
-                
-                <div class="q-pa-none" style="max-height: 128px; overflow-y: auto;">
-                  <q-list dense class="q-pa-none">
-                    <q-item
-                      v-for="media in mediaFiles"
-                      :key="media.id"
-                      dense
-                      clickable
-                      class="q-pa-xs rounded hover:bg-grey-2 cursor-pointer"
-                      @click="previewMedia(media)"
-                    >
-                      <q-item-section avatar class="min-width-auto q-pr-xs">
-                        <q-icon name="image" size="14px" color="blue-5" />
-                      </q-item-section>
-                      <q-item-section>
-                        <q-item-label class="text-caption text-weight-medium text-grey-9" lines="1">
-                          {{ media.filename || media.label }}
-                        </q-item-label>
-                        <q-item-label caption class="text-grey-6">
-                          {{ media.filetype || 'Unknown' }} • {{ formatFileSize(media.filesize) }}
-                        </q-item-label>
-                      </q-item-section>
-                    </q-item>
-                  </q-list>
-                </div>
-              </q-expansion-item>
+              <ReferenceList
+                v-model:expanded="referencesExpanded"
+                :references="references"
+                @add="showReferenceSearch = true"
+                @select="openReference"
+              />
+              <MediaList
+                v-model:expanded="mediaFilesExpanded"
+                :media="mediaFiles"
+                @select="previewMedia"
+              />
             </q-list>
           </div>
         </div>
@@ -512,6 +435,8 @@ import FileActionDialog from '../components/project/FileActionDialog.vue'
 import FileUpload from '../components/ui/FileUpload.vue'
 import PubMedSearch from '../components/ui/PubMedSearch.vue'
 import ReferenceSearchDialog from '../components/project/ReferenceSearchDialog.vue'
+import ReferenceList from '../components/project/ReferenceList.vue'
+import MediaList from '../components/project/MediaList.vue'
 import { useReferenceStore } from '../stores/reference'
 import { useMediaStore } from '../stores/media'
 import { useUserStore } from '../stores/user'
@@ -623,14 +548,6 @@ async function handleUploadMedia() {
 
 async function uploadMediaFiles(files) {
   await mediaStore.upload({ projectId: projectId.value, files })
-}
-
-function formatFileSize(bytes) {
-  if (bytes === 0) return '0 B'
-  if (!bytes) return ''
-  const sizes = ['B', 'KB', 'MB', 'GB', 'TB']
-  const i = Math.floor(Math.log(bytes) / Math.log(1024))
-  return `${(bytes / Math.pow(1024, i)).toFixed(1)} ${sizes[i]}`
 }
 
 function openReference(ref) {
